NCL-1875: Enable reading GRIB2 files > 2GB Replaced the g2clib seekgb routine with an internal version that supports large files. The new routine is called _g2_seekgb.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15030 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/NclGRIB2.h
@@ -456,7 +456,7 @@ typedef struct _g2Sec7 {
 } G2Sec7;
 
 typedef struct  _g2Rec {
-    off_t offset;
+    size_t offset;
     int rec_size;
     int version;        /* GRIB version */
     char *table_source_name;
@@ -634,7 +634,7 @@ struct _Grib2AttInqRecList {
 
 
 struct _Grib2RecordInqRec {
-    off_t offset;
+    size_t offset;
     int field_num;
     int rec_size;
     int rec_num;
